Calculate next build number in a trigger instead ruby
This is an attempt to fix #120. When calculating next build number in
Ruby, there is a good chance that between the time we get last number
and insert it to a new build, the same will be done with other build.
Using SQL does not have that problem.

This is not ideal solution in terms of complexity, but other solutions
are hard to implement with ActiveRecord:

  * we could use INSERT statement with SELECT, for example:

        INSERT INTO builds (number) VALUES
          ( SELECT MAX(number) + 1 FROM builds WHERE repository_id = 1 )

    but it's hard to change how INSERTs are done in AR (also we would
    need to disable prepared statements for the time of doing this
    query)

  * we could catch ActiveRecord::RecordNotUnique error, but if it's
    thrown in a transaction, it will rollback it, which will prevent
    retries

Overally trigger seems to be the simplest option as it does not need any
changes in application code and it will work in any place we need to
create a build.